introduce a logical-atomicity-style pmem specification
There is an adapter that takes the existing spec, PersistentMemoryRegion,
and implements a logical-atomicity spec, PersistentMemoryRegionV2,
on top of it.  This helps ensure that the logically-atomic spec doesn't
say anything unsound in terms of PCM resources.  However, the adapter
would have to be part of the TCB for purposes of crash safety reasoning.